Finalisation API Simulator
- Passage des ID à Long (Permet de le set à null car c'est un Object et non un type primitif).

- Suppression de tous les liens de Barrack et Truck (Model + Base de donnée)

- Suppression du Seeder

+++ src/main/java/com/projettransversal/api/Repositories/IncidentRepository.java
@@ -8,10 +8,10 @@ import org.springframework.stereotype.Repository;
 import java.util.Collection;
 
 @Repository
-public interface IncidentRepository extends CrudRepository<Incident, Integer> {
+public interface IncidentRepository extends CrudRepository<Incident, Long> {
 
     @Query(
             value = "select * from incident where intensity = ?1 and incident_type = ?2 and map_item_id = ?3",
             nativeQuery = true)
-    Collection<Incident> findByData(float intensity, String y, int mapItemId);
+    Collection<Incident> findByData(float intensity, String y, Long mapItemId);
 }
