+++ src/main/java/com/projettransversal/api/Services/IServices/ICrudService.java
@@ -5,7 +5,10 @@ import java.util.Optional;
 
 public interface ICrudService<T> {
     List<T> findAll();
-    Optional<T> findById(int id);
+
+    Optional<T> findById(Long id);
+
     T insertOrUpdate(T obj);
+
     void delete(T obj);
 }
